Runtime/http: stupid interface
git-svn-id: http://www.virtualbox.org/svn/vbox/trunk@45340 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Runtime/common/misc/http.cpp
@@ -203,7 +203,7 @@ RTR3DECL(int) RTHttpCertDigest(RTHTTP hHttp, char *pcszCert, size_t cbCert,
                                uint8_t **pabSha1,   size_t *pcbSha1,
                                uint8_t **pabSha512, size_t *pcbSha512)
 {
-    BIO *cert = BIO_new_mem_buf(pcszCert, cbCert);
+    BIO *cert = BIO_new_mem_buf(pcszCert, (int)cbCert);
     if (!cert)
         return VERR_INTERNAL_ERROR;
